Moved to pytest-style fixture and added implementation parameters
This might turn out to be a bad idea, since we actually test all
possible combinations in all possible positions across the
implementations. But it is what we want right now.

Notice that having Eclair as a recipient for a connection does not
work since we have no good way of retrieving its ID which is required
for the connection. It'll just complain about an invalid MAC.